+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 - **breaking:** Introduced a new property in theme called `tabs_active_highlight`, default color is `#ff00ff`.
 - **breaking:** Removed breadcrumb navigation.
 - **breaking:** Introduced a new property in theme called `bar`, default color changed is `#1b1a1a`.
-- **breaking:** `BottomTab` is now default for all platforms.
+- **breaking:** `CollapsedTab` is now default for all platforms.
 - Introduce `navigation.hide-if-single` property (Ref: [#595](https://github.com/raphamorim/rio/issues/595)).
 - Performance update: Render repeated styled fragments as one rect.
 - Sugarloaf API has changed from `Sugar` primitives to `Content`.